Add: Model files for each table.
